+++ examples/fastify-http2.ts
@@ -2,7 +2,7 @@
 import * as fs from 'fs';
 import { join } from 'path';
 
-import fastify from 'fastify';
+import { fastify } from 'fastify';
 
 import { FileSystemStorage } from '../src/send-stream';
 
@@ -21,7 +21,7 @@ const storage = new FileSystemStorage(join(__dirname, 'assets'));
 app.route({
 	method: ['HEAD', 'GET'],
 	url: '*',
-	handler: async ({ req }, { res }) => {
+	handler: async ({ raw: req }, { raw: res }) => {
 		const result = await storage.prepareResponse(req.url, req);
 		result.send(res);
 	},