+++ examples/pushstate-server.ts
@@ -20,7 +20,7 @@ app.use(async (req, res, next) => {
 				['', 'index.html'],
 				req,
 				// if the mime type can be determined from path then this is probably an error so add 404 status
-				storage.mimeModule.getType(pathParts[pathParts.length - 1])
+				storage.mimeTypesCharset(pathParts[pathParts.length - 1])
 					? { statusCode: 404 }
 					: {},
 			);